additional improvements per review
