+++ lib/rx/client.rb
@@ -3,7 +3,7 @@ require 'faraday'
 require 'multi_json'
 require 'common/client/errors'
 require 'common/client/concerns/client_methods'
-require 'common/client/concerns/session_based_client'
+require 'common/client/concerns/mhv_session_based_client'
 require 'common/client/middleware/response/json_parser'
 require 'common/client/middleware/response/raise_error'
 require 'common/client/middleware/response/snakecase'
@@ -16,15 +16,9 @@ require 'rx/api/sessions'
 module Rx
   # Core class responsible for api interface operations
   class Client
-    include Common::ClientMethods
-    include Common::SessionBasedClient
+    include Common::Client::MHVSessionBasedClient
     include Rx::API::Prescriptions
     include Rx::API::Sessions
-
-    def initialize(session:)
-      @config = Rx::Configuration.instance
-      @session = Rx::ClientSession.find_or_build(session)
-    end
 
     private
